# Application/Backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Flux Backend",
    "version": "1.0.0",
    "description": "A comprehensive API management system for environment files, web flows, and file uploads",
    "contact": {
      "name": "API Flux Team",
      "email": "support@apiflux.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.apiflux.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/web-flows/import": {
      "post": {
        "summary": "Create a web flow from imported file data",
        "tags": [
          "WebFlows"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "The name of the web flow",
                    "example": "My Imported Web Flow"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the web flow",
                    "example": "A web flow imported from a JSON file"
                  },
                  "basePath": {
                    "type": "string",
                    "description": "Base path for the web flow",
                    "example": "/api/v1"
                  },
                  "isFolder": {
                    "type": "boolean",
                    "description": "Whether this is a folder",
                    "default": false
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Array of tags",
                    "example": [
                      "api",
                      "imported"
                    ]
                  },
                  "playgroundConfig": {
                    "type": "object",
                    "properties": {
                      "nodes": {
                        "type": "array",
                        "items": {
                          "type": "object"
                        },
                        "description": "Array of flow nodes"
                      },
                      "edges": {
                        "type": "array",
                        "items": {
                          "type": "object"
                        },
                        "description": "Array of flow edges"
                      }
                    },
                    "required": [
                      "nodes",
                      "edges"
                    ]
                  },
                  "parentId": {
                    "type": "integer",
                    "nullable": true,
                    "description": "Parent folder ID"
                  }
                },
                "required": [
                  "name",
                  "playgroundConfig"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Web flow created from import successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Web flow created from import successfully"
                    },
                    "webFlow": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "basePath": {
                          "type": "string"
                        },
                        "isFolder": {
                          "type": "boolean"
                        },
                        "tags": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "config": {
                          "type": "object",
                          "properties": {
                            "nodes": {
                              "type": "array"
                            },
                            "edges": {
                              "type": "array"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid playground configuration. Must contain nodes and edges arrays"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}